all rare words in wiktionary
A term or sense that is attested but not used commonly either in spoken or written language, even less so than uncommon terms. A rare term or sense is typically difficult to find even when deliberately searching for it, and may be attested only a handful of times in accessible corpora.